requirements.txt
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
setuptools==69.5.1

